:bug: proper JSON Pointer escape in diagnostic messages
