+++ include/nlohmann/json.hpp
@@ -8651,7 +8651,7 @@ class basic_json
                 for (auto it = source.cbegin(); it != source.cend(); ++it)
                 {
                     // escape the key name to be used in a JSON patch
-                    const auto key = json_pointer::escape(it.key());
+                    const auto key = detail::escape(it.key());
 
                     if (target.find(it.key()) != target.end())
                     {
@@ -8675,7 +8675,7 @@ class basic_json
                     if (source.find(it.key()) == source.end())
                     {
                         // found a key that is not in this -> add it
-                        const auto key = json_pointer::escape(it.key());
+                        const auto key = detail::escape(it.key());
                         result.push_back(
                         {
                             {"op", "add"}, {"path", path + "/" + key},
